Use `class_type_is_data_frame()` in `is_data_frame()`
