+++ src/utils-dispatch.c
@@ -91,17 +91,6 @@ enum vctrs_class_type class_type_impl(r_obj* class) {
   return VCTRS_CLASS_unknown;
 }
 
-bool class_type_is_data_frame(enum vctrs_class_type type) {
-  switch (type) {
-  case VCTRS_CLASS_data_frame:
-  case VCTRS_CLASS_bare_data_frame:
-  case VCTRS_CLASS_bare_tibble:
-    return true;
-  default:
-    return false;
-  }
-}
-
 static
 const char* class_type_as_str(enum vctrs_class_type type) {
   switch (type) {
